[tracy] move `tracy` under `tools/`
The motivation for this change is to enable easier downstream usage of
`tracy`, by consolidating various scripts and functionalities used
during profiling in one place.

Majority of the touched files are either moved or have trivial changes.
However, there are important changes to the wheel creation/editable
installation.

`tracy` is added as a top-level package in `setup.py`. This works for
wheels, but the editable installation is more complicated.

Previously, editable installation enabled resolving `import ttnn` by
creating `easy-install.pth` file in the `site-packages` directory. This
file contained the path to `ttnn` package, which is automatically added
to `sys.path` by python. However, now that we have two packages, i
haven't found a way to add a second path there - hence, `python -m
tracy` couldn't work, while `python -m tools.tracy` worked, since it
could be resolved from the current working directory.

This problem prompted me to try using latest `pip`, since they modified
how the editable installations work. New version solved the problem for
`tracy`, however it created a new problem for `ttnn` - `import ttnn` was
now broken.

I couldn't find a proper solution, so i've implemented a
workaround in `setup.py` (look there for more details).

No functional changes are intended.

NOTE: also contains small fix for the release workflow.
