+++ models/demos/llama3_70b_galaxy/tests/test_decoder_device_perf.py
@@ -14,7 +14,7 @@ from models.demos.llama3_70b_galaxy.tt.llama_common import (
 )
 from models.perf.benchmarking_utils import BenchmarkData, BenchmarkProfiler
 from models.perf.device_perf_utils import run_device_perf
-from tt_metal.tools.profiler.process_model_log import (
+from tracy.process_model_log import (
     get_latest_ops_log_filename,
 )
