+++ models/demos/llama3_70b_galaxy/tests/test_prefill_device_perf.py
@@ -8,7 +8,7 @@ import pandas as pd
 from collections import defaultdict
 from models.perf.benchmarking_utils import BenchmarkData, BenchmarkProfiler
 from models.perf.device_perf_utils import run_device_perf
-from tt_metal.tools.profiler.process_model_log import (
+from tracy.process_model_log import (
     get_latest_ops_log_filename,
 )
