+++ tech_reports/YoloV4-TTNN/yolov4.md
@@ -128,7 +128,7 @@ The diagram below illustrates that using HEIGHT_SHARDED for the first convolutio
 #### How to generate the graph?
 1. Generate the Performance Sheet:
    - First, build the profiler using the command: ./scripts/build_scripts/build_with_profiler_opt.sh.
-   - Next, execute the command: ./tt_metal/tools/profiler/profile_this.py -n <Folder_name> -c "pytest <path_to_test_file>".
+   - Next, execute the command: ./tools/tracy/profile_this.py -n <Folder_name> -c "pytest <path_to_test_file>".
    - Download the generated CSV file (the file path will be displayed in the terminal).
 2. Access the Analysis Tool:
    - Go throught this [comment](https://github.com/tenstorrent/tt-metal/issues/12468#issuecomment-2341711534).
