+++ tests/tt_metal/microbenchmarks/ethernet/test_all_ethernet_links_bandwidth.py
@@ -13,7 +13,7 @@ from tests.tt_metal.microbenchmarks.ethernet.test_all_ethernet_links_common impo
 
 from models.utility_functions import is_single_chip
 
-from tt_metal.tools.profiler.common import PROFILER_LOGS_DIR, PROFILER_DEVICE_SIDE_LOG
+from tracy.common import PROFILER_LOGS_DIR, PROFILER_DEVICE_SIDE_LOG
 
 profiler_log_path = PROFILER_LOGS_DIR / PROFILER_DEVICE_SIDE_LOG
