+++ tests/ttnn/unit_tests/operations/ccl/perf/run_reduce_scatter_profile.sh
@@ -59,7 +59,7 @@ done
 
 # Function to run the profiling command and extract the CSV path
 run_profile_and_extract_csv() {
-    command="./tt_metal/tools/profiler/profile_this.py -n reduce_scatter_$TARGET -c 'pytest tests/ttnn/unit_tests/operations/ccl/perf/test_ccl_perf.py::test_reduce_scatter_on_$TARGET'"
+    command="./tools/tracy/profile_this.py -n reduce_scatter_$TARGET -c 'pytest tests/ttnn/unit_tests/operations/ccl/perf/test_ccl_perf.py::test_reduce_scatter_on_$TARGET'"
 
     if [ "$DEBUG" = true ]; then
         echo "Running profiling command for target $TARGET in debug mode..."
