+++ tools/tracy/process_device_log.py
@@ -17,8 +17,8 @@ import seaborn as sns
 import click
 from loguru import logger
 
-from tt_metal.tools.profiler.common import PROFILER_ARTIFACTS_DIR
-import tt_metal.tools.profiler.device_post_proc_config as device_post_proc_config
+from tracy.common import PROFILER_ARTIFACTS_DIR
+import tracy.device_post_proc_config as device_post_proc_config
 
 SUM_MARKER_ID_START = 3000
