+++ tools/tracy/process_ops_logs.py
@@ -19,9 +19,8 @@ from collections import deque
 import click
 from loguru import logger
 
-from tt_metal.tools.profiler.process_device_log import import_log_run_stats
-import tt_metal.tools.profiler.device_post_proc_config as device_post_proc_config
-from tt_metal.tools.profiler.common import (
+from tracy.process_device_log import import_log_run_stats
+from tracy.common import (
     PROFILER_DEVICE_SIDE_LOG,
     PROFILER_HOST_SIDE_LOG,
     PROFILER_ARTIFACTS_DIR,
@@ -32,6 +31,7 @@ from tt_metal.tools.profiler.common import (
     generate_logs_folder,
     generate_reports_folder,
 )
+from tracy import device_post_proc_config
 
 yaml.SafeDumper.ignore_aliases = lambda *args: True
